+++ internal/rubyserver/testhelper_test.go
@@ -1,0 +1,22 @@
+package rubyserver
+
+import (
+	"os"
+	"testing"
+
+	"gitlab.com/gitlab-org/gitaly/internal/testhelper"
+)
+
+var (
+	testRepo = testhelper.TestRepository()
+)
+
+func TestMain(m *testing.M) {
+	os.Exit(testMain(m))
+}
+
+func testMain(m *testing.M) int {
+	defer testhelper.MustHaveNoChildProcess()
+
+	return m.Run()
+}