+++ ruby/lib/gitaly_server.rb
@@ -2,6 +2,7 @@ require 'gitaly'
 
 require_relative 'gitlab/git.rb'
 
+require_relative 'gitaly_server/client.rb'
 require_relative 'gitaly_server/utils.rb'
 require_relative 'gitaly_server/commit_service.rb'
 require_relative 'gitaly_server/diff_service.rb'
@@ -14,6 +15,7 @@ module GitalyServer
   REPO_PATH_HEADER = 'gitaly-repo-path'.freeze
   GL_REPOSITORY_HEADER = 'gitaly-gl-repository'.freeze
   REPO_ALT_DIRS_HEADER = 'gitaly-repo-alt-dirs'.freeze
+  GITALY_SERVERS_HEADER = 'gitaly-servers'.freeze
 
   def self.repo_path(call)
     call.metadata.fetch(REPO_PATH_HEADER)
@@ -25,6 +27,10 @@ module GitalyServer
 
   def self.repo_alt_dirs(call)
     call.metadata.fetch(REPO_ALT_DIRS_HEADER)
+  end
+
+  def self.client(call)
+    Client.new(call.metadata[GITALY_SERVERS_HEADER])
   end
 
   def self.register_handlers(server)