+++ ruby/lib/gitaly_server/repository_service.rb
@@ -17,5 +17,15 @@ module GitalyServer
 
       Gitaly::HasLocalBranchesResponse.new(value: repo.has_local_branches?)
     end
+
+    def fetch_source_branch(request, call)
+      bridge_exceptions do
+        source_repository = Gitlab::Git::GitalyRemoteRepository.new(request.source_repository, call)
+        repository = Gitlab::Git::Repository.from_gitaly(request.repository, call)
+        result = repository.fetch_source_branch!(source_repository, request.source_branch, request.target_ref)
+
+        Gitaly::FetchSourceBranchResponse.new(result: result)
+      end
+    end
   end
 end